\name{affected_norms}

\alias{affected_norms}


\title{ Affected Norms }

\description{
  Computes a list containing all seperate norms used in the ruli
}

\usage{
affected_norms(data)
}


\arguments{
  \item{data}{ Single-row dataframe with the the column "Norm" }
}

\value{
  Returns a list containing all seperate norms used in the ruling.
}

\references{ Austrian law information system \url{http://www.ris.bka.gv.at/} }

\author{ Martin Kober <martin.kober@wu-wien.ac.at>, Georg M. Zajko <georg.zajko@wu-wien.ac.at> }

\keyword{misc}

