\name{put_RIS_jurisdictions_into_data_frame}

\alias{put_RIS_jurisdictions_into_data_frame}

\title{Internal RIS function, converts output from read\_RIS\_jurisdiction 
to a dataframe}

\description{
  Takes a list of output vectors from read\_RIS\_jurisdiction and converts them to a 
  dataframe with a layout suitable for databse insertion.}

\usage{
	put_RIS_jurisdictions_into_data_frame(out)
}

\arguments{
  \item{      out}{ A list of character vectors, as generated by read\_RIS\_jurisdiction. }
}

\value{
  Returns a data.frame with all columns used in the database in the correct order.
}

\references{ Austrian law information system \url{http://www.ris.bka.gv.at/} }

\author{ Kurt Hornik <kurt.hornik@wu-wien.ac.at>, Martin Kober <martin.kober@wu-wien.ac.at>, Georg M. Zajko <georg.zajko@wu-wien.ac.at> }

\examples{
\dontrun{
x = lapply(100:105, read_RIS_jurisdiction)
y = put_RIS_jurisdictions_into_data_frame(x)
y[["Geschaeftszahl"]]
}
}

\keyword{ misc }
