\name{senate_number}

\alias{senate_number}


\title{ Senate Number }

\description{
  Converts a string from the database into the senate number.
}


\usage{
senate_number(data)
}


\arguments{
  \item{data}{ Single-row dataframe with the the column "Geschaeftszahl" }
}


\value{
  Returns the number of the senate handling the case.
}


\references{ Austrian law information system \url{http://www.ris.bka.gv.at/} }

\author{ Martin Kober <martin.kober@wu-wien.ac.at>, Georg M. Zajko <georg.zajko@wu-wien.ac.at> }

\keyword{misc}

