\name{senate_size}

\alias{senate_size}


\title{ Senate Size }

\description{
  Determines the number of judges on the case.
}

\usage{
senate_size(data)
}


\arguments{
  \item{data}{ Single-row dataframe with the the column "Betreff" }
}

\value{
  Returns the number of judges working on the case.
}

\references{ Austrian law information system \url{http://www.ris.bka.gv.at/} }

\author{ Martin Kober <martin.kober@wu-wien.ac.at>, Georg M. Zajko <georg.zajko@wu-wien.ac.at> }


\keyword{misc}

